Add boss spawn

+++ src/compass.rs
@@ -1,7 +1,7 @@
 use bevy::{prelude::*, render::mesh::PrimitiveTopology};
-use rand::Rng;
+// use rand::Rng;
 
-use crate::{boss::BOSS_Z, spaceship::Spaceship, WINDOW_HEIGHT, WINDOW_WIDTH};
+use crate::{boss::BossCore, spaceship::Spaceship, WINDOW_HEIGHT, WINDOW_WIDTH};
 
 const COMPASS_POSITION: Vec3 = Vec3 {
     x: WINDOW_WIDTH / 2.0 - 150.0,
@@ -15,7 +15,7 @@ const NEEDLE_POSITION: Vec3 = Vec3 {
     y: 0.0,
     z: 0.0,
 };
-const DISTANCE_TO_TARGET: f32 = 5000.0;
+// const DISTANCE_TO_TARGET: f32 = 5000.0;
 const FONT: &'static str = "fonts/FiraSans-Bold.ttf";
 const FONT_SIZE: f32 = 20.0;
 const COLOR: Color = Color::DARK_GRAY;
@@ -28,9 +28,8 @@ const NEEDLE_SCALE: f32 = 0.13;
 // }
 
 #[derive(Component)]
-pub struct Compass {
-    target: Vec3,
-}
+pub struct Compass;
+// { pub target: Vec3, }
 
 #[derive(Component)]
 pub struct Needle;
@@ -40,8 +39,14 @@ pub fn setup(
     mut meshes: ResMut<Assets<Mesh>>,
     mut materials: ResMut<Assets<ColorMaterial>>,
     asset_server: Res<AssetServer>,
-    query: Query<&Transform, With<Camera>>,
+    query_camera: Query<&Transform, With<Camera>>,
+    query_boss: Query<&Transform, With<BossCore>>,
+    query_spaceship: Query<&Transform, With<Spaceship>>,
 ) {
+    let camera = query_camera.single();
+    let boss = query_boss.single();
+    let spaceship = query_spaceship.single();
+
     let font = asset_server.load(FONT);
     let text_style = TextStyle {
         font,
@@ -49,12 +54,6 @@ pub fn setup(
         color: COLOR,
     };
     let text_alignment = TextAlignment::CENTER_LEFT;
-
-    let mut rng = rand::thread_rng();
-    let camera = query.single();
-    let x = rng.gen_range(-DISTANCE_TO_TARGET..DISTANCE_TO_TARGET);
-    let y_max = (DISTANCE_TO_TARGET.powi(2) - x.powi(2)).sqrt();
-    let y = rng.gen_range(-y_max..y_max);
 
     let c1 = Vec3::new(75.0, 0.0, 0.0);
     let c2 = Vec3::new(-50.0, 50.0, 0.0);
@@ -65,10 +64,10 @@ pub fn setup(
     mesh.insert_attribute(Mesh::ATTRIBUTE_POSITION, v_pos);
 
     let compass = commands
-        .spawn(Compass {
-            target: Vec3::new(x, y, BOSS_Z),
-            // boss_spawned: false,
-        })
+        .spawn(Compass)
+        // { target: Vec3::new(x, y, BOSS_Z),
+        // boss_spawned: false,
+        // })
         .insert(Text2dBundle {
             text: Text::from_section("", text_style).with_alignment(text_alignment),
             transform: Transform::from_translation(camera.translation + COMPASS_POSITION),
@@ -81,7 +80,13 @@ pub fn setup(
         .insert(ColorMesh2dBundle {
             mesh: meshes.add(mesh).into(),
             transform: Transform::from_translation(NEEDLE_POSITION)
-                .with_scale(Vec3::splat(NEEDLE_SCALE)),
+                .with_scale(Vec3::splat(NEEDLE_SCALE))
+                .with_rotation(Quat::from_rotation_arc_2d(
+                    Vec2::X,
+                    (boss.translation - spaceship.translation)
+                        .truncate()
+                        .normalize(),
+                )),
             material: materials.add(COLOR.into()),
             ..default()
         })
@@ -91,23 +96,26 @@ pub fn setup(
 }
 
 pub fn update(
-    mut query_compass: Query<(&Compass, &mut Transform, &mut Text)>,
+    mut query_compass: Query<(&mut Transform, &mut Text), With<Compass>>,
     mut query_needle: Query<&mut Transform, (With<Needle>, Without<Compass>)>,
     query_spaceship: Query<&Transform, (With<Spaceship>, Without<Compass>, Without<Needle>)>,
+    query_boss: Query<&Transform, (With<BossCore>, Without<Compass>, Without<Needle>)>,
     query_camera: Query<&Transform, (With<Camera>, Without<Compass>, Without<Needle>)>,
 ) {
     if let Ok(spaceship) = query_spaceship.get_single() {
-        let camera = query_camera.single();
-        let (compass, mut compass_transform, mut text) = query_compass.single_mut();
-        let mut needle = query_needle.single_mut();
-        compass_transform.translation = camera.translation + COMPASS_POSITION;
-        needle.rotation = Quat::from_rotation_arc_2d(
-            Vec2::X,
-            (compass.target - spaceship.translation)
-                .truncate()
-                .normalize(),
-        );
-        let distance = (compass.target - spaceship.translation).length();
-        text.sections[0].value = format!("Target: {:<7.0}", distance);
+        if let Ok(boss) = query_boss.get_single() {
+            let camera = query_camera.single();
+            let (mut compass, mut text) = query_compass.single_mut();
+            let mut needle = query_needle.single_mut();
+            compass.translation = camera.translation + COMPASS_POSITION;
+            needle.rotation = Quat::from_rotation_arc_2d(
+                Vec2::X,
+                (boss.translation - spaceship.translation)
+                    .truncate()
+                    .normalize(),
+            );
+            let distance = (boss.translation - spaceship.translation).length();
+            text.sections[0].value = format!("Target: {:<7.0}", distance);
+        }
     }
 }
